+++ src/Appwrite/Utopia/Response/Model/AttributeLine.php
@@ -12,14 +12,18 @@ class AttributeLine extends Attribute
 
         $this
             ->addRule('default', [
-                'type' => self::TYPE_JSON,
+                'type' => self::TYPE_ARRAY,
                 'description' => 'Default value for attribute when not provided. Cannot be set when attribute is required.',
                 'default' => null,
                 'required' => false,
-                'example' => '[[0, 0], [1, 1]]'
+                'example' => [[0, 0], [1, 1]]
             ])
         ;
     }
+
+    public array $conditions = [
+        'type' => 'linestring',
+    ];
 
     /**
      * Get Name