+++ src/Appwrite/Utopia/Response/Model/AttributePoint.php
@@ -12,14 +12,18 @@ class AttributePoint extends Attribute
 
         $this
             ->addRule('default', [
-                'type' => self::TYPE_JSON,
+                'type' => self::TYPE_ARRAY,
                 'description' => 'Default value for attribute when not provided. Cannot be set when attribute is required.',
                 'default' => null,
                 'required' => false,
-                'example' => '[0, 0]'
+                'example' => [0, 0]
             ])
         ;
     }
+
+    public array $conditions = [
+        'type' => 'point',
+    ];
 
     /**
      * Get Name