+++ tests/e2e/Services/Databases/Legacy/DatabasesBase.php
@@ -6537,7 +6537,7 @@ trait DatabasesBase
             'x-appwrite-key' => $this->getProject()['apiKey']
         ]), [
             'required' => false,
-            'default' => json_encode([[0, 0], [1, 1]]),
+            'default' => [[0, 0], [1, 1]],
         ]);
 
         $this->assertEquals(200, $response['headers']['status-code']);
@@ -6564,7 +6564,7 @@ trait DatabasesBase
             'x-appwrite-project' => $this->getProject()['$id'],
             'x-appwrite-key' => $this->getProject()['apiKey']
         ]), [
-            'default' => json_encode([0, 0]),
+            'default' => [0, 0],
             'required' => false
         ]);
 
@@ -6769,7 +6769,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::contains('lineAttr', [[1.0, 1.0]])->toString()]
+            'queries' => [Query::contains('lineAttr', [[1.1, 1.1]])->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertCount(1, $response['body']['documents']);
@@ -6790,7 +6790,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::intersects('polyAttr', [[5.0, 5.0]])->toString()]
+            'queries' => [Query::intersects('polyAttr', [5.0, 5.0])->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(1, $response['body']['total']);
@@ -6801,7 +6801,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::notIntersects('polyAttr', [[60.0, 60.0]])->toString()]
+            'queries' => [Query::notIntersects('polyAttr', [60.0, 60.0])->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(3, $response['body']['total']);
@@ -6812,7 +6812,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::overlaps('polyAttr', [$overlapPoly])->toString()]
+            'queries' => [Query::overlaps('polyAttr', $overlapPoly)->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(1, $response['body']['total']);
@@ -6824,7 +6824,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::notOverlaps('polyAttr', [$noOverlapPoly])->toString()]
+            'queries' => [Query::notOverlaps('polyAttr', $noOverlapPoly)->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(3, $response['body']['total']);
@@ -6876,7 +6876,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::crosses('lineAttr', [$crossLine])->toString()]
+            'queries' => [Query::crosses('lineAttr', $crossLine)->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(1, $response['body']['total']);
@@ -6888,7 +6888,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::notCrosses('lineAttr', [$nonCrossLine])->toString()]
+            'queries' => [Query::notCrosses('lineAttr', $nonCrossLine)->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(3, $response['body']['total']);
@@ -6899,7 +6899,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::touches('polyAttr', [$touchPoly])->toString()]
+            'queries' => [Query::touches('polyAttr', $touchPoly)->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(2, $response['body']['total']);
@@ -6911,7 +6911,7 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'queries' => [Query::notTouches('polyAttr', [$farPoly])->toString()]
+            'queries' => [Query::notTouches('polyAttr', $farPoly)->toString()]
         ]);
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals(3, $response['body']['total']);
@@ -7634,19 +7634,6 @@ trait DatabasesBase
             'attributes' => ['pOptional'],
         ]);
         $this->assertEquals(202, $retriedIndex['headers']['status-code']);
-
-        // Passing orders to spatial index should not throw error(in case of mariadb)
-        $ordersIndex = $this->client->call(Client::METHOD_POST, '/databases/' . $databaseId . '/collections/' . $collectionId . '/indexes', array_merge([
-            'content-type' => 'application/json',
-            'x-appwrite-project' => $this->getProject()['$id'],
-            'x-appwrite-key' => $this->getProject()['apiKey']
-        ]), [
-            'key' => 'idx_required_point_with_orders',
-            'type' => Database::INDEX_SPATIAL,
-            'attributes' => ['pRequired'],
-            'orders' => ['ASC']
-        ]);
-        $this->assertEquals(202, $ordersIndex['headers']['status-code']);
 
         // Cleanup
         $this->client->call(Client::METHOD_DELETE, '/databases/' . $databaseId . '/collections/' . $collectionId, array_merge([